weather_service/requirements.txt
# API and Data Processing
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.3

# Kafka
kafka-python==2.0.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.12.0
pylint==3.0.3
mypy==1.7.1

# Logging and Error Handling
retry-requests==2.0.0

# Development Tools
setuptools>=65.5.1
wheel>=0.40.0
pip>=23.0.1